refactor(spec): split InAppNotificationSubstance into forms, content, and data hook
